refactor: export device type from sdk
