+++ packages/hms-video-web/src/device-manager/DeviceManager.ts
@@ -2,8 +2,7 @@ import { DeviceStorageManager } from './DeviceStorage';
 import AnalyticsEventFactory from '../analytics/AnalyticsEventFactory';
 import { HMSException } from '../error/HMSException';
 import { EventBus } from '../events/EventBus';
-import { HMSDeviceChangeEvent } from '../interfaces';
-import type { DeviceMap } from '../interfaces/HMSDeviceManager';
+import { DeviceMap, DeviceType, HMSDeviceChangeEvent } from '../interfaces';
 import { HMSAudioTrackSettingsBuilder, HMSVideoTrackSettingsBuilder } from '../media/settings';
 import { HMSLocalAudioTrack, HMSLocalTrack, HMSLocalVideoTrack } from '../media/tracks';
 import { IStore } from '../sdk/store';
@@ -11,9 +10,9 @@ import HMSLogger from '../utils/logger';
 import { debounce } from '../utils/timer-utils';
 
 export type SelectedDevices = {
-  audioInput?: MediaDeviceInfo;
-  videoInput?: MediaDeviceInfo;
-  audioOutput?: MediaDeviceInfo;
+  [DeviceType.audioInput]?: MediaDeviceInfo;
+  [DeviceType.videoInput]?: MediaDeviceInfo;
+  [DeviceType.audioOutput]?: MediaDeviceInfo;
 };
 
 type DeviceAndGroup = Partial<MediaTrackSettings>;
