--- packages/hms-video-web/src/interfaces/HMSDeviceManager.ts
@@ -1,5 +1,0 @@
-export interface DeviceMap {
-  audioInput: MediaDeviceInfo[];
-  audioOutput: MediaDeviceInfo[];
-  videoInput: MediaDeviceInfo[];
-}
