--- packages/hms-video-web/src/interfaces/device-change-event.ts
@@ -1,9 +1,0 @@
-import { DeviceMap } from './HMSDeviceManager';
-import { HMSException } from '../error/HMSException';
-
-export interface HMSDeviceChangeEvent {
-  error?: HMSException;
-  devices: DeviceMap;
-  selection?: MediaDeviceInfo;
-  type: 'audioOutput' | 'audioInput' | 'video';
-}
