--- packages/hms-video-web/src/interfaces/device-change-listener.ts
@@ -1,5 +1,0 @@
-import { HMSDeviceChangeEvent } from './device-change-event';
-
-export interface DeviceChangeListener {
-  onDeviceChange?(event: HMSDeviceChangeEvent): void;
-}
