--- packages/hms-video-web/src/interfaces/device.ts
@@ -1,10 +1,0 @@
-export default interface HMSDevice {
-  videoInputs: HMSVideoDevice[];
-  videoOutputs: HMSVideoDevice[];
-  audioInputs: HMSAudioDevice[];
-  audioOutputs: HMSAudioDevice[];
-}
-
-type HMSVideoDevice = HMSDevice;
-
-type HMSAudioDevice = HMSDevice;
