+++ packages/hms-video-web/src/interfaces/index.ts
@@ -10,9 +10,8 @@ export * from './change-track-state';
 export * from './leave-room-request';
 export type { HMSSimulcastLayerDefinition, SimulcastLayer, RID, HMSPreferredSimulcastLayer } from './simulcast-layers';
 export { HMSSimulcastLayer, simulcastMapping } from './simulcast-layers';
-export type { DeviceMap } from './HMSDeviceManager';
 export * from './track-settings';
-export * from './device-change-event';
+export * from './devices';
 export * from './playlist';
 export * from './hms';
 export * from './rtmp-recording-config';
