+++ packages/react-sdk/src/hooks/useDevices.ts
@@ -1,14 +1,13 @@
 import { useCallback } from 'react';
-import { selectDevices, selectIsAllowedToPublish, selectLocalMediaSettings } from '@100mslive/hms-video-store';
+import {
+  DeviceType,
+  selectDevices,
+  selectIsAllowedToPublish,
+  selectLocalMediaSettings,
+} from '@100mslive/hms-video-store';
 import { hooksErrHandler } from '../hooks/types';
 import { useHMSActions, useHMSStore } from '../primitives/HmsRoomProvider';
 import { logErrorHandler } from '../utils/commons';
-
-export enum DeviceType {
-  videoInput = 'videoInput',
-  audioInput = 'audioInput',
-  audioOutput = 'audioOutput',
-}
 
 export type DeviceTypeAndInfo<T> = {
   [key in DeviceType]?: T;
